+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/NumberOfPositionsCriterionTest.java
@@ -35,13 +35,15 @@ import org.ta4j.core.BaseTradingRecord;
 import org.ta4j.core.Position;
 import org.ta4j.core.Trade;
 import org.ta4j.core.TradingRecord;
+import org.ta4j.core.AnalysisCriterion.PositionFilter;
 import org.ta4j.core.mocks.MockBarSeries;
 import org.ta4j.core.num.Num;
 
 public class NumberOfPositionsCriterionTest extends AbstractCriterionTest {
 
     public NumberOfPositionsCriterionTest(Function<Number, Num> numFunction) {
-        super((params) -> new NumberOfPositionsCriterion(), numFunction);
+        super((params) -> params.length == 0 ? new NumberOfPositionsCriterion()
+                : new NumberOfPositionsCriterion((PositionFilter) params[0]), numFunction);
     }
 
     @Test
@@ -50,6 +52,8 @@ public class NumberOfPositionsCriterionTest extends AbstractCriterionTest {
 
         AnalysisCriterion buyAndHold = getCriterion();
         assertNumEquals(0, buyAndHold.calculate(series, new BaseTradingRecord()));
+        assertNumEquals(0, getCriterion(PositionFilter.PROFIT).calculate(series, new BaseTradingRecord()));
+        assertNumEquals(0, getCriterion(PositionFilter.LOSS).calculate(series, new BaseTradingRecord()));
     }
 
     @Test
@@ -60,6 +64,11 @@ public class NumberOfPositionsCriterionTest extends AbstractCriterionTest {
 
         AnalysisCriterion buyAndHold = getCriterion();
         assertNumEquals(2, buyAndHold.calculate(series, tradingRecord));
+        assertNumEquals(2, getCriterion(PositionFilter.PROFIT).calculate(series, tradingRecord));
+
+        TradingRecord tradingRecordLoss = new BaseTradingRecord(Trade.buyAt(1, series), Trade.sellAt(3, series),
+                Trade.buyAt(3, series), Trade.sellAt(4, series));
+        assertNumEquals(2, getCriterion(PositionFilter.LOSS).calculate(series, tradingRecordLoss));
     }
 
     @Test
@@ -69,6 +78,25 @@ public class NumberOfPositionsCriterionTest extends AbstractCriterionTest {
         AnalysisCriterion positionsCriterion = getCriterion();
 
         assertNumEquals(1, positionsCriterion.calculate(series, position));
+
+        Position positionProfit = new Position(Trade.buyAt(0, series), Trade.sellAt(2, series));
+        assertNumEquals(1, getCriterion(PositionFilter.PROFIT).calculate(series, positionProfit));
+
+        Position positionLoss = new Position(Trade.buyAt(1, series), Trade.sellAt(3, series));
+        assertNumEquals(1, getCriterion(PositionFilter.LOSS).calculate(series, positionLoss));
+    }
+
+    @Test
+    public void calculateWithTwoShortPositions() {
+        MockBarSeries seriesProfit = new MockBarSeries(numFunction, 110, 105, 110, 100, 95, 105);
+        TradingRecord tradingRecordProfit = new BaseTradingRecord(Trade.sellAt(0, seriesProfit),
+                Trade.buyAt(1, seriesProfit), Trade.sellAt(2, seriesProfit), Trade.buyAt(4, seriesProfit));
+        assertNumEquals(2, getCriterion(PositionFilter.PROFIT).calculate(seriesProfit, tradingRecordProfit));
+
+        MockBarSeries seriesLoss = new MockBarSeries(numFunction, 100, 105, 110, 100, 95, 105);
+        TradingRecord tradingRecord = new BaseTradingRecord(Trade.sellAt(0, seriesLoss), Trade.buyAt(1, seriesLoss),
+                Trade.sellAt(3, seriesLoss), Trade.buyAt(5, seriesLoss));
+        assertNumEquals(2, getCriterion().calculate(seriesLoss, tradingRecord));
     }
 
     @Test
@@ -76,5 +104,21 @@ public class NumberOfPositionsCriterionTest extends AbstractCriterionTest {
         AnalysisCriterion criterion = getCriterion();
         assertTrue(criterion.betterThan(numOf(3), numOf(6)));
         assertFalse(criterion.betterThan(numOf(7), numOf(4)));
+
+        AnalysisCriterion criterionProfit = getCriterion(PositionFilter.PROFIT);
+        assertTrue(criterionProfit.betterThan(numOf(6), numOf(3)));
+        assertFalse(criterionProfit.betterThan(numOf(4), numOf(7)));
+
+        AnalysisCriterion criterionLoss = getCriterion(PositionFilter.LOSS);
+        assertTrue(criterionLoss.betterThan(numOf(3), numOf(6)));
+        assertFalse(criterionLoss.betterThan(numOf(7), numOf(4)));
+    }
+
+    @Test
+    public void testCalculateOneOpenPositionShouldReturnZero() {
+        openedPositionUtils.testCalculateOneOpenPositionShouldReturnExpectedValue(numFunction,
+                getCriterion(PositionFilter.PROFIT), 0);
+        openedPositionUtils.testCalculateOneOpenPositionShouldReturnExpectedValue(numFunction,
+                getCriterion(PositionFilter.LOSS), 0);
     }
 }