+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/pnl/NetCriterionTest.java
@@ -31,6 +31,7 @@ import java.util.function.Function;
 
 import org.junit.Test;
 import org.ta4j.core.AnalysisCriterion;
+import org.ta4j.core.AnalysisCriterion.PositionFilter;
 import org.ta4j.core.BaseTradingRecord;
 import org.ta4j.core.Trade;
 import org.ta4j.core.TradingRecord;
@@ -38,10 +39,10 @@ import org.ta4j.core.analysis.criteria.AbstractCriterionTest;
 import org.ta4j.core.mocks.MockBarSeries;
 import org.ta4j.core.num.Num;
 
-public class NetLossCriterionTest extends AbstractCriterionTest {
+public class NetCriterionTest extends AbstractCriterionTest {
 
-    public NetLossCriterionTest(Function<Number, Num> numFunction) {
-        super((params) -> new NetLossCriterion(), numFunction);
+    public NetCriterionTest(Function<Number, Num> numFunction) {
+        super((params) -> new NetCriterion((PositionFilter) params[0]), numFunction);
     }
 
     @Test
@@ -50,8 +51,21 @@ public class NetLossCriterionTest extends AbstractCriterionTest {
         TradingRecord tradingRecord = new BaseTradingRecord(Trade.buyAt(0, series), Trade.sellAt(2, series),
                 Trade.buyAt(3, series), Trade.sellAt(5, series));
 
-        AnalysisCriterion loss = getCriterion();
+        AnalysisCriterion profit = getCriterion(PositionFilter.PROFIT);
+        assertNumEquals(15, profit.calculate(series, tradingRecord));
+
+        AnalysisCriterion loss = getCriterion(PositionFilter.LOSS);
         assertNumEquals(0, loss.calculate(series, tradingRecord));
+    }
+
+    @Test
+    public void calculateOnlyWithProfitPositions2() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 105, 100, 80, 85, 120);
+        TradingRecord tradingRecord = new BaseTradingRecord(Trade.buyAt(0, series), Trade.sellAt(1, series),
+                Trade.buyAt(2, series), Trade.sellAt(5, series));
+
+        AnalysisCriterion profit = getCriterion(PositionFilter.PROFIT);
+        assertNumEquals(25, profit.calculate(series, tradingRecord));
     }
 
     @Test
@@ -60,7 +74,7 @@ public class NetLossCriterionTest extends AbstractCriterionTest {
         TradingRecord tradingRecord = new BaseTradingRecord(Trade.buyAt(0, series), Trade.sellAt(1, series),
                 Trade.buyAt(2, series), Trade.sellAt(5, series));
 
-        AnalysisCriterion loss = getCriterion();
+        AnalysisCriterion loss = getCriterion(PositionFilter.LOSS);
         assertNumEquals(-35, loss.calculate(series, tradingRecord));
     }
 
@@ -70,19 +84,29 @@ public class NetLossCriterionTest extends AbstractCriterionTest {
         TradingRecord tradingRecord = new BaseTradingRecord(Trade.sellAt(0, series), Trade.buyAt(1, series),
                 Trade.sellAt(2, series), Trade.buyAt(5, series));
 
-        AnalysisCriterion loss = getCriterion();
+        AnalysisCriterion profit = getCriterion(PositionFilter.PROFIT);
+        assertNumEquals(0, profit.calculate(series, tradingRecord));
+
+        AnalysisCriterion loss = getCriterion(PositionFilter.LOSS);
         assertNumEquals(-35, loss.calculate(series, tradingRecord));
     }
 
     @Test
     public void betterThan() {
-        AnalysisCriterion criterion = getCriterion();
-        assertTrue(criterion.betterThan(numOf(2.0), numOf(1.5)));
-        assertFalse(criterion.betterThan(numOf(1.5), numOf(2.0)));
+        AnalysisCriterion criterionProfit = getCriterion(PositionFilter.PROFIT);
+        assertTrue(criterionProfit.betterThan(numOf(2.0), numOf(1.5)));
+        assertFalse(criterionProfit.betterThan(numOf(1.5), numOf(2.0)));
+
+        AnalysisCriterion criterionLoss = getCriterion(PositionFilter.LOSS);
+        assertTrue(criterionLoss.betterThan(numOf(2.0), numOf(1.5)));
+        assertFalse(criterionLoss.betterThan(numOf(1.5), numOf(2.0)));
     }
 
     @Test
     public void testCalculateOneOpenPositionShouldReturnZero() {
-        openedPositionUtils.testCalculateOneOpenPositionShouldReturnExpectedValue(numFunction, getCriterion(), 0);
+        openedPositionUtils.testCalculateOneOpenPositionShouldReturnExpectedValue(numFunction,
+                getCriterion(PositionFilter.PROFIT), 0);
+        openedPositionUtils.testCalculateOneOpenPositionShouldReturnExpectedValue(numFunction,
+                getCriterion(PositionFilter.LOSS), 0);
     }
 }